requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
debugpy==1.8.1
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.2
flake8==7.0.0
future==1.0.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
orjson==3.9.13
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyflakes==3.2.0
pyserial==3.5
pytest==8.0.0
python-dotenv==1.0.1
python-multipart==0.0.8
PyYAML==6.0.1
serial==0.0.97
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
